requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
halo==0.0.31
halp==1.0.0
idna==3.10
Jinja2==3.1.6
keyboard==0.13.5
log-symbols==0.0.14
MarkupSafe==3.0.3
narwhals==2.7.0
networkx==3.4.2
packaging==25.0
plotly==6.3.1
ply==3.11
psutil==7.1.2
pyasn1==0.6.1
pycryptodomex==3.23.0
pygame==2.6.1
pysmi==1.6.2
pysnmp==7.1.8
PyYAML==6.0.3
requests==2.32.5
semantic-version==2.10.0
setuptools-rust==1.12.0
six==1.17.0
spinners==0.0.24
termcolor==3.2.0
tqdm==4.67.1
urllib3==2.5.0
yara-python==4.5.4
